Add comments sensative looking material for usage.
This change simply adds comments to secrets and other password/key
material that are used in testing. The usage of these secrets is okay
(they're all either dummy tokens or auth material set for local
instances spun up during tests), but adding comments to provide some
background for anyone browsing the code.

Also redacted a cert used in the buildah sample. While this was
safe since it was generated specifically for the sample, there isn't a strict
need for a valid cert here, so we can replace this with something that looks
cert-like instead.